Moving DeleteComponet into ReportComponet
